Created aliasNamespaceEx.cpp and h_AliasNamespaceEx.hpp
These files show an example of using 'Namespace Aliases' which helps us to
abstract the chains of access, for example:
We have the main namespace: 'Company'
The second namespace inside Company : 'Department'
The third namespace inside Departmen : 'Project'
This third project nested in Department contains inside a class called 'Project'.
To access the class members we can use an alias for the long chain needed to access them.
For example: original chain will be Company::Department::Project;
declaring an alias for this chain will be: namespace projectAlias = Company::Department::Project;
And Lastly, we create an object from this Project class like this using the alias:
	aliasProject::Project myProject1("charlie project:);
Here we are using the constructor of the Project class and creating an object called 'myProject1'
Finally, to acces it's public methods, simply write the object name and the the (.) operator to call it:
	myProject1.printProjectInfo();	// asumming it has a method called printProjectInfo();