Created anonymousNamespaceEx2.cpp to demonstrate the use of anonymous namespace
This example showcase using an anonymous namespace with a helper function. The function defined inside the
anonymous namespace can be accessed inside the main() function. However, this helper function and other members that
can be added inside this anonymous namespace, are only accessible for this specific filed, it cannot be referenced in another file.
Lastly to used or to reference a member of this anonymous namespace, we use the name of the member without an accompanying namespace name, becuase it does not have a name!